+++ src/components/SupplierDetails.js
@@ -1,12 +1,20 @@
 import { useSuppliersContext } from "../hooks/useSuppliersContext";
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useAuthContext } from "../hooks/useAuthContext.js"
 
 const SupplierDetails = ({ supplier }) => {
   const { dispatchSuppliers } = useSuppliersContext()
+  const { user } = useAuthContext()
   
   const handleClickSupplier = async () => {
+    if (!user) {
+      return
+    }
     const response = await fetch('/api/suppliers/' + supplier._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
